+++ build.gradle
@@ -56,7 +56,7 @@ dependencies {
 	implementation group: 'io.jsonwebtoken', name: 'jjwt-jackson', version: '0.11.5'
 
 	//feign
-	implementation 'org.springframework.cloud:spring-cloud-starter-openfeign:4.1.0'
+	implementation 'org.springframework.cloud:spring-cloud-starter-openfeign:2.14.0'
 
 	//oauth
 	implementation 'org.springframework.boot:spring-boot-starter-oauth2-client'
@@ -79,5 +79,6 @@ dependencies {
 	implementation 'net.logstash.logback:logstash-logback-encoder:7.4'
 
 	//s3
-	implementation 'org.springframework.cloud:spring-cloud-starter-aws:2.2.6.RELEASE'
+	implementation("software.amazon.awssdk:bom:2.21.0")
+	implementation("software.amazon.awssdk:s3:2.21.0")
 }
