+++ src/main/java/com/permitseoul/permitserver/domain/admin/util/aws/S3Service.java
@@ -1,0 +1,4 @@
+package com.permitseoul.permitserver.domain.admin.util.aws;
+
+public class S3Service {
+}
