+++ src/main/java/com/permitseoul/permitserver/domain/admin/util/aws/config/S3Config.java
@@ -1,0 +1,4 @@
+package com.permitseoul.permitserver.domain.admin.util.aws.config;
+
+public class S3Config {
+}
