+++ src/main/java/com/permitseoul/permitserver/global/aop/resolver/timetableblock/TimetableBlockPathVariableResolver.java
@@ -1,5 +1,4 @@
-package com.permitseoul.permitserver.global.resolver.event;
-
+package com.permitseoul.permitserver.global.aop.resolver.timetableblock;
 
 import com.permitseoul.permitserver.global.exception.UrlSecureException;
 import com.permitseoul.permitserver.global.response.code.ErrorCode;
@@ -18,13 +17,13 @@ import java.util.Map;
 
 @Component
 @RequiredArgsConstructor
-public class EventIdPathVariableResolver implements HandlerMethodArgumentResolver {
-    private static final String EVENT_ID_PATH_VARIABLE = "eventId";
+public class TimetableBlockPathVariableResolver implements  HandlerMethodArgumentResolver{
+    private static final String BLOCK_ID_PATH_VARIABLE = "blockId";
     private final SecureUrlUtil secureUrlUtil;
 
     @Override
     public boolean supportsParameter(MethodParameter parameter) {
-        return parameter.hasParameterAnnotation(EventIdPathVariable.class);
+        return parameter.hasParameterAnnotation(TimetableBlockIdPathVariable.class);
     }
 
     @Override
@@ -32,11 +31,12 @@ public class EventIdPathVariableResolver implements HandlerMethodArgumentResolve
         final HttpServletRequest request = (HttpServletRequest) webRequest.getNativeRequest();
         final Map<String, String> pathVariables = (Map<String, String>) request.getAttribute(HandlerMapping.URI_TEMPLATE_VARIABLES_ATTRIBUTE);
 
-        final String eventId = pathVariables.get(EVENT_ID_PATH_VARIABLE);
+        final String blockId = pathVariables.get(BLOCK_ID_PATH_VARIABLE);
         try {
-            return secureUrlUtil.decode(eventId);
+            return secureUrlUtil.decode(blockId);
         } catch (Exception e) {
             throw new UrlSecureException(ErrorCode.BAD_REQUEST_ID_DECODE_ERROR);
         }
     }
 }
+
